// src/api-generated/model/turma-basica-dto.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * API Horários Acadêmicos
 * Documentação da API para o Sistema de Elaboração de Horário e Atribuição de Disciplinas
 * OpenAPI spec version: 1.0
 */
import type { DisciplinaOfertadaBasicaDto } from './disciplina-ofertada-basica-dto';
import type { ProfessorBasicoDto } from './professor-basico-dto';

export interface TurmaBasicaDto {
  /** ID da turma */
  id: string;
  /** Código da turma */
  codigoDaTurma: string;
  /** Dados da disciplina ofertada */
  disciplinaOfertada: DisciplinaOfertadaBasicaDto;
  /** Professor alocado à turma */
  professorAlocado?: ProfessorBasicoDto;
}